interceptor 역할 분할
