+++ frontend/src/components/utility/interceptors/interceptors.js
@@ -1,32 +1,72 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import {tokenRefresh} from '../auth.js';
+import axios from 'axios';
+export default function setInterceptors(instance) {
+    instance.interceptors.request.use(
+        async (res) => {
+            const access = await AsyncStorage.getItem('accessToken');
+            const refresh = await AsyncStorage.getItem('refreshToken');
+    
+            if (access) {
+                res.headers['Authorization'] = `Bearer ${access}`;
+            }
+    
+            const newConfig = {...res};
+            newConfig.url = `${res.url}`;
+    
+            return newConfig;
+        }
+    );
 
-export default function withoutInterceptors(instance) {
     instance.interceptors.response.use((response) => {
         return response
-    }, 
-    async function(error) {
+    }, async function(error) {
         const originalRequest = error.config;
-        console.log('step #1############');
-
-        if (error.response.status === 401 && !originalRequest.retry) {
-            const accessToken = AsyncStorage.getItem('accessToken');
-            const refreshToken = AsyncStorage.getItem('refreshToken');
-
-            const newData = await tokenRefresh(accessToken, refreshToken);
-
-            if (newData.success) {
-                originalRequest.retry = true;
-                AsyncStorage.setItem('accessToken', newData.access);
-
-                instance.defaults.headers['Authorization'] = `Bearer ${accessToken}`;
-                return instance(originalRequest);
-            } else {
-                return error;
+            console.log('step #1############');
+            console.log(error.config);
+            console.log('retry', originalRequest.retry);
+            if (error.response.status === 401 && !originalRequest.retry) {
+                const accessToken = await AsyncStorage.getItem('accessToken');
+                const refreshToken = await AsyncStorage.getItem('refreshToken');
+                console.log('access', accessToken);
+                console.log('refresh', refreshToken);
+    
+                const newData = await tokenRefresh(accessToken, refreshToken);
+                console.log('newData', newData);
+    
+                if (newData.status === 'New Access Token granted') {
+                    originalRequest.retry = true;
+                    AsyncStorage.setItem('accessToken', newData.tokens.access);
+                    const regex = /"data":{.+}/
+                    console.log('regex', regex.exec(originalRequest.data));
+                    console.log('original Request', originalRequest.data);
+    
+                    const content_regex = /"content":".+?"/
+                    const latitude_regex = /"latitude":[0-9\.]+/
+                    const longitude_regex = /"longitude":[0-9\.]+/
+                    console.log(content_regex.exec(originalRequest.data)[0].split(':')[1].slice(1,-1));
+                    console.log(latitude_regex.exec(originalRequest.data)[0].split(':')[1]);
+                    console.log(longitude_regex.exec(originalRequest.data)[0].split(':')[1]);
+                    const originData = {
+                        content:content_regex.exec(originalRequest.data)[0].split(':')[1].slice(1,-1),
+                        latitude:latitude_regex.exec(originalRequest.data)[0].split(':')[1],
+                        longitude:longitude_regex.exec(originalRequest.data)[0].split(':')[1]
+                    }
+    
+    
+                    originalRequest.headers['Authorization'] = `Bearer ${newData.tokens.access}`;
+                    originalRequest.data = originData;
+                    console.log(originalRequest.data);
+                    console.log('original', originalRequest)
+                    return axios(originalRequest);
+                } else {
+                    return error;
+                }
             }
-        }
-        return error;
-    })
+            return Promise.reject(error);
+        })
+
+        return instance;
 }
